Fixed a bug with the HttpRequester
The application was crashing on the test http request sometimes. It
appears that all request must be run in another thread, because the Main
Thread (UI Thread) freezes. I've removed the HttpRequester and put all
the login in a HttpActivity which has tasks that work on a different
thread. Every activity must extend this activity in order to have the
defined tasks. The freeze problem came from the onPostExecute() which is
executed on the UI thread and there was network connection going in it.
I've created a ResponsePair class which holds only the response data
from the server (in json string format) and the status code of the
response. With the status code I can decide which event to call (either
onSuccess, or onError)