Custom click event on the Register button added
It took me some time to understand how to prevent a Dialog from
dismissing when either the negative or the positive buttons are pressed.
This was I can make client validation for the user input and show a
Toast for incorrect input. Then I will just have to call the dismiss()
method of the dialog.