Main fragments made reusable
The fragments were working only with the current user's id. That was
they are not reusable. I made a constructor that takes the userId as
parameter but there was an error that I must not user constructor rather
than default once for all Fragments, and that the preferred way is with
fragment.setArguments(Bundle) and fragment.getArguments(). Now the code
is more coupled, but that's android development I guess :D