Add Skill implemented
The adding of a skill works fine the the update of the UI is hard. I was
trying to update it for the last 2.5 hours and I am not getting
anywhere. I am using Communicator pattern to communicate between the
different fragments, but when I get the properties of the other fragment
from the activity I get NullPointerException, but the fragment is
visualized (I don't get it).