EditProfileActivity Set-Up-User-Photo optimization
When the user chooses an image from the camera or from the gallery,
after the redirection the image is compressed and assigned to the
"short[] userPhotoUnsignedByteArray" field in another thread. This way
the UI is not freezing if the picture is very big and the time to
compress is bigger. When the user presses the Edit button, there is only
initialization of the editModel that will be send to the server an no
hard processing like the compression of the photo, which was moved to
the background processes in other threads.