SearchResults ListView implemented
There is a strange bug. if a user A has a photo X (AX) and user B has a
photo Y (BY) - the photos can actually be reverted in the list view -
(AY) and (BX). This is because all of the photos are handles in other
threads in order to prevent UI freeze if the photos are too big, but the
references of the sub-vies (imageView and textView) are the changing
constantly. I will look it through and will find a solution.