namespaces, layouts and classes renamed (not all)
I will need separate fragments for the UserActivity. For example: I have
AboutFragment in the wiredup.fragments.profile.activity and I will
create one more AboutFragment in the wiredup.fragments.user.activity
they will have slightly different content.

I also renamed part of the adapters because I will need new adapters for
the new fragments. For example: I have
ProfileActivityCertificatesAdapter which handles the certificates in the
profile activity certificate fragment. This adapter does a lot of
things: can red, add, delete certificates.
I will need UserActivityCertificatesAdapter which will handle only
reading of activities. The problem is that the previous adapter is not
very reusable as I thought, but it will be hard to make it reusable.
This happened maybe because I am skill a rookie. It's not such a big
deal but... :(