Modifications to the MessagesFragment made
When I delete a message from the ALL messages tab, this message is not
being deleted in the other 2 tabs. That's because once the messages are
get from the database, they are cached in the memory. Now I removed the
cache, but unfortunately every time the user changes the TAB an http
request is sent to the database.